securanet_backend/requirements.txt
# Core Web Framework


# REST API support
djangorestframework>=3.14.0

# File Monitoring
watchdog>=3.0.0

# File Type Analysis (e.g., mime types)
python-magic>=0.4.27
python-magic-bin>=0.4.14  # for Windows systems

# AI / Machine Learning support (optional but useful)
scikit-learn>=1.3.0
pandas>=2.2.0
numpy>=1.24.0

# Environment variable management
python-decouple>=3.8

# Enhanced development shell
ipython>=8.20.0

click==8.1.3
pyyaml==6.0.1

dnspython>=2.4.0
requests>=2.31.0
scikit-learn>=1.3.0

joblib>=1.3.0 
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0

filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3

packaging==25.0
pillow==11.0.0

regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
setuptools==70.2.0
sqlparse==0.5.3
sympy==1.13.3
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0